change in the location of the temples
