+++ carpa/urls.py
@@ -23,6 +23,7 @@ urlpatterns = [
     path('', include('main.urls')),
     path('', include('reports.urls')),
     path('', include('harvests.urls')),
+    path('', include('sources.urls')),
     path('admin/', admin.site.urls),
 ]
