+++ harvests/urls.py
@@ -1,6 +1,6 @@
 # general imports
 from django.urls import path
-from harvests.views import people, vehicles, origin, harvests
+from harvests.views import people, harvests
 
 # api imports
 
@@ -10,6 +10,4 @@ from harvests.views import people, vehicles, origin, harvests
 urlpatterns = [
     path('harvests', harvests, name="harvests"),
     path('people', people, name="people"),
-    path('vehicles', vehicles, name="vehicles"),
-    path('origin', origin, name="origin"),
 ]
