+++ harvests/views.py
@@ -7,9 +7,4 @@ def harvests(request):
 def people(request):   
     return render(request, "people.html")
 
-def vehicles(request):   
-    return render(request, "vehicles.html")
-
-def origin(request):   
-    return render(request, "origin.html")
